% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Money-class.R
\docType{class}
\name{Money-class}
\alias{Money-class}
\alias{Money}
\title{An S4 class to represent Money}
\description{
An S4 class to represent Money
}
\section{Slots}{

\describe{
\item{\code{currency}}{A three letter currency symbol}

\item{\code{value}}{A numeric value corresponding to money}

\item{\code{format}}{A character code corresponding to available format}
}}

