% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/money-package.R
\docType{package}
\name{money-package}
\alias{money-package}
\title{money: A fun package providing money class}
\description{
The package provides money class that can be used to print currency in formats
  that will appeal to MS Excel aficionados!
}
